+++ packages/cli/src/commands/add.ts
@@ -79,6 +79,7 @@ export async function add(key: string, options: AddOptions) {
   if (!options.skipCodegen) {
     generate({
       config: options.config,
+      descriptorsPackage: options.descriptorsPackage,
     })
   }
 }