+++ packages/cli/src/commands/commonOptions.ts
@@ -1,4 +1,5 @@
 export interface CommonOptions {
   config?: string
   skipCodegen?: boolean
+  descriptorsPackage?: boolean
 }