+++ packages/cli/src/commands/ink.ts
@@ -40,6 +40,7 @@ export const ink = {
     if (!options.skipCodegen) {
       generate({
         config: options.config,
+        descriptorsPackage: options.descriptorsPackage,
       })
     }
   },
@@ -63,6 +64,7 @@ export const ink = {
     if (!options.skipCodegen) {
       generate({
         config: options.config,
+        descriptorsPackage: options.descriptorsPackage,
       })
     }
   },