+++ packages/cli/src/commands/remove.ts
@@ -18,6 +18,7 @@ export async function remove(key: string, options: CommonOptions) {
   if (!options.skipCodegen) {
     generate({
       config: options.config,
+      descriptorsPackage: options.descriptorsPackage,
     })
   }
 }