+++ packages/cli/src/commands/update.ts
@@ -51,6 +51,7 @@ export async function update(
     console.log(`Updating descriptors`)
     await generate({
       config: options.config,
+      descriptorsPackage: options.descriptorsPackage,
     })
   }