+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Option flag `--no-descriptors-package` to generate descriptors without installing them as a package.
+
 ## 0.9.17 - 2024-11-07
 
 - Patch dependencies (smoldot)